backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.30
pydantic==2.7.4
pydantic-settings==2.3.4
python-jose==3.3.0
# passlib[bcrypt]==1.7.4
python-multipart==0.0.9
passlib[argon2]==1.7.4
argon2-cffi==23.1.0

